docker_compose/simple_project/app/requirements.txt
Django==4.2.11
psycopg[binary]==3.1.18
python-dotenv==1.0.1
split-settings==1.0.0
gunicorn==23.0.0
django-debug-toolbar==5.2.0
django-cors-headers==4.7.0